requirements.txt
torch>=2.0.0
transformers>=4.51.0
numpy>=1.24.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
python-dotenv>=1.0.0
pymilvus>=2.3.0
anthropic>=0.7.0
tika>=2.6.0
opencv-python>=4.8.0
torchvision>=0.15.0
torchaudio>=2.0.0
llama-index>=0.8.0
requests>=2.31.0
PyGithub>=2.1.1
msal>=1.25.0
Pillow>=10.0.0
scikit-learn>=1.3.0
openai-whisper>=20231117
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
jinja2>=3.1.2
aiofiles>=23.1.0
fastapi==0.115.12
starlette==0.46.1
langchain_openai==0.3.12
anthropic==0.49.0
# External dependencies
# ffmpeg - install via: brew install ffmpeg (macOS) or apt-get install ffmpeg (Linux)